refactor: change filename api-interfaces.ts
